requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116
torch==1.13.1+cu116
torchvision==0.14.1+cu116
torchaudio==0.13.1
pillow
progress
pyyaml
tensorboard
scipy
pykalman
six